[QFEED-158] Refactor: 관리자 질문 API AdminController로 이동
